backend/requirements.txt
aiomysql==0.3.2
annotated-types==0.7.0
anyio==4.12.0
apscheduler==3.11.0
bcrypt==3.2.2
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
fastapi==0.125.0
feedparser==6.0.11
greenlet==3.3.0
h11==0.16.0
httptools==0.7.1
httpx==0.28.1
idna==3.11
passlib==1.7.4
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyMySQL==1.1.2
python-dotenv==1.2.1
PyYAML==6.0.3
SQLAlchemy==2.0.45
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
faiss-cpu==1.11.0
numpy>=1.24.0
openai>=1.35.0
